v2.4 Update
Merged a pull request, bug fixes, and deleted all the code for advanced
features rendering this the "free" version of Google Maps for
ExpressionEngine. All features in the previous v2.3 beta have been
merged into the v3 release. Commercial release is coming soon with all
new advanced features and a complete rewrite of the code based.