v2.5 Update
Version 2.x has officially come to a close. I am only leaving it
available for those that need a free solution. This version will not be
officially supported. Everyone is encouraged to upgrade to v3. I put A
LOT of time and hard work into v3 and is something I am extremely proud
of. Thanks for everyone who supported this add-on, and I hope you
continue to support v3.

I also updated the download package with the original documentation and
examples, as they are no longer available online.